requirements.txt
# Multi-Index RAG Requirements
# Install with: pip install -r requirements.txt

# Core ML
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# LLM
google-genai>=1.0.0
langchain-google-genai

# LangGraph Pipeline
langgraph>=0.2.0
langsmith>=0.1.0
langchain-core>=0.2.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0

# Evaluation
scikit-learn>=1.3.0

# Development
pytest>=7.0.0

